requirements.txt
streamlit==1.28.0
google-generativeai==0.3.1
PyMuPDF==1.23.0
python-docx==0.8.11
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0
spacy==3.7.2
nltk==3.8.1
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4